+++ tests/testdata/check_lua/blocking_io.py
@@ -1,0 +1,6 @@
+times = [
+    # <block check-lua="tests/testdata/check_lua/blocking_io.lua" check-lua-timeout="1">
+    'morning',
+    'afternoon',
+    # </block>
+]